+++ test_runner/src/main/kotlin/ftl/shard/Shard.kt
@@ -23,12 +23,12 @@ data class TestShard(
 )
 
 /** List of shards containing test names as a string. */
-typealias StringShards = MutableList<MutableList<String>>
+typealias StringShards = List<MutableList<String>>
 
 fun List<TestShard>.stringShards(): StringShards {
     return this.map { shard ->
         shard.testMethods.map { it.name }.toMutableList()
-    }.toMutableList()
+    }.toList()
 }
 
 /*